sweetalert 추가

+++ src/app/_global/components/FileImages.tsx
@@ -4,6 +4,7 @@ import styled from 'styled-components'
 import Image from 'next/image'
 import { FaRegWindowClose } from 'react-icons/fa'
 import LayerPopup from './LayerPopup'
+import useConfirmDialog from '../hooks/useConfirmDialog'
 import useFetchCSR from '../hooks/useFetchCSR'
 import color from '../styles/color'
 
@@ -58,22 +59,27 @@ const ImageItem = ({
   const { seq, fileUrl, thumbBaseUrl, fileName, image } = item
   const [open, setOpen] = useState<boolean>(false)
   const fetchCSR = useFetchCSR()
-
+  const confirmDialog = useConfirmDialog()
   const onClose = useCallback(() => setOpen(false), [])
   const onShow = useCallback(() => setOpen(true), [])
 
   const onRemove = useCallback(
     (seq) => {
-      fetchCSR(`/file/delete/${seq}`, { method: 'DELETE' })
-        .then((res) => res.json())
-        .then((item) => {
-          // 삭제 후 후속 처리
-          if (typeof callback === 'function') {
-            callback(item)
-          }
-        })
+      confirmDialog({
+        text: '정말 삭제하시겠습니까?',
+        confirmCallback: () => {
+          fetchCSR(`/file/delete/${seq}`, { method: 'DELETE' })
+            .then((res) => res.json())
+            .then((item) => {
+              // 삭제 후 후속 처리
+              if (typeof callback === 'function') {
+                callback(item)
+              }
+            })
+        },
+      })
     },
-    [fetchCSR, callback],
+    [fetchCSR, callback, confirmDialog],
   )
 
   return (
